Update (#30)
New graphs for classes, functions, and files. The implementation has been moved to a separate grapher package.

New outputs to the console. The implementation has been moved to the separate package representator. Added the ability to receive data in json format.

Added a server for working with the analyzer from other programs.

Various minor fixes and improvements.